refactor: split oversized modules into focused sub-packages
- Split sensornames.py (1296 lines) into sensornames/ package
  (__init__, controls, diagnostics, energy, ev, ml)
- Split hourly_data_populator.py (961 lines) into package
  (__init__, consumption, prices_solcast)
- Extract conversion.py, ha_helpers.py, time_windows.py from misc.py
- Fix all import paths across the codebase
- Update test imports to match new module structure
