+++ custom_components/hsem/custom_sensors/house_consumption_power_sensor.py
@@ -39,12 +39,10 @@ from custom_components.hsem.custom_sensors.utility_meter_sensor import (
     HSEMUtilityMeterSensor,
 )
 from custom_components.hsem.entity import HSEMEntity
+from custom_components.hsem.utils.conversion import convert_to_float
+from custom_components.hsem.utils.ha_helpers import ha_get_entity_state_and_convert
 from custom_components.hsem.utils.logger import HSEM_LOGGER as _LOGGER
-from custom_components.hsem.utils.misc import (
-    convert_to_float,
-    get_config_value,
-    ha_get_entity_state_and_convert,
-)
+from custom_components.hsem.utils.misc import get_config_value
 from custom_components.hsem.utils.sensornames import (
     get_energy_average_sensor_entity_id,
     get_energy_average_sensor_name,
