+++ custom_components/hsem/utils/sensornames/__init__.py
@@ -1,0 +1,377 @@
+"""Sensor name, unique ID, and entity ID generators for HSEM entities.
+
+This package splits the original sensornames.py into domain-specific modules.
+All public names are re-exported here so existing call sites continue to work
+with ``from custom_components.hsem.utils.sensornames import X``.
+"""
+
+from custom_components.hsem.utils.sensornames.controls import (
+    get_batteries_schedule_1_switch_entity_id,
+    get_batteries_schedule_1_switch_key,
+    get_batteries_schedule_1_switch_name,
+    get_batteries_schedule_1_switch_unique_id,
+    get_batteries_schedule_2_switch_entity_id,
+    get_batteries_schedule_2_switch_key,
+    get_batteries_schedule_2_switch_name,
+    get_batteries_schedule_2_switch_unique_id,
+    get_batteries_schedule_3_switch_entity_id,
+    get_batteries_schedule_3_switch_key,
+    get_batteries_schedule_3_switch_name,
+    get_batteries_schedule_3_switch_unique_id,
+    get_charge_efficiency_number_entity_id,
+    get_charge_efficiency_number_key,
+    get_charge_efficiency_number_name,
+    get_charge_efficiency_number_unique_id,
+    get_discharge_efficiency_number_entity_id,
+    get_discharge_efficiency_number_key,
+    get_discharge_efficiency_number_name,
+    get_discharge_efficiency_number_unique_id,
+    get_extended_attributes_switch_entity_id,
+    get_extended_attributes_switch_key,
+    get_extended_attributes_switch_name,
+    get_extended_attributes_switch_unique_id,
+    get_read_only_switch_entity_id,
+    get_read_only_switch_key,
+    get_read_only_switch_name,
+    get_read_only_switch_unique_id,
+    get_schedule_1_end_time_entity_id,
+    get_schedule_1_end_time_key,
+    get_schedule_1_end_time_name,
+    get_schedule_1_end_time_unique_id,
+    get_schedule_1_start_time_entity_id,
+    get_schedule_1_start_time_key,
+    get_schedule_1_start_time_name,
+    get_schedule_1_start_time_unique_id,
+    get_schedule_2_end_time_entity_id,
+    get_schedule_2_end_time_key,
+    get_schedule_2_end_time_name,
+    get_schedule_2_end_time_unique_id,
+    get_schedule_2_start_time_entity_id,
+    get_schedule_2_start_time_key,
+    get_schedule_2_start_time_name,
+    get_schedule_2_start_time_unique_id,
+    get_schedule_3_end_time_entity_id,
+    get_schedule_3_end_time_key,
+    get_schedule_3_end_time_name,
+    get_schedule_3_end_time_unique_id,
+    get_schedule_3_start_time_entity_id,
+    get_schedule_3_start_time_key,
+    get_schedule_3_start_time_name,
+    get_schedule_3_start_time_unique_id,
+    get_verbose_logging_switch_entity_id,
+    get_verbose_logging_switch_key,
+    get_verbose_logging_switch_name,
+    get_verbose_logging_switch_unique_id,
+)
+from custom_components.hsem.utils.sensornames.diagnostics import (
+    get_applier_status_sensor_entity_id,
+    get_applier_status_sensor_name,
+    get_applier_status_sensor_unique_id,
+    get_battery_soc_sensor_entity_id,
+    get_battery_soc_sensor_name,
+    get_battery_soc_sensor_unique_id,
+    get_daily_plan_vs_actual_sensor_entity_id,
+    get_daily_plan_vs_actual_sensor_key,
+    get_daily_plan_vs_actual_sensor_name,
+    get_daily_plan_vs_actual_sensor_unique_id,
+    get_degraded_mode_sensor_entity_id,
+    get_degraded_mode_sensor_name,
+    get_degraded_mode_sensor_unique_id,
+    get_force_mode_sensor_entity_id,
+    get_force_mode_sensor_name,
+    get_force_mode_sensor_unique_id,
+    get_force_working_mode_selector_entity_id,
+    get_force_working_mode_selector_key,
+    get_force_working_mode_selector_name,
+    get_force_working_mode_selector_unique_id,
+    get_forecast_accuracy_sensor_entity_id,
+    get_forecast_accuracy_sensor_name,
+    get_forecast_accuracy_sensor_unique_id,
+    get_hardware_writes_sensor_entity_id,
+    get_hardware_writes_sensor_name,
+    get_hardware_writes_sensor_unique_id,
+    get_last_updated_sensor_entity_id,
+    get_last_updated_sensor_name,
+    get_last_updated_sensor_unique_id,
+    get_missing_entities_sensor_entity_id,
+    get_missing_entities_sensor_name,
+    get_missing_entities_sensor_unique_id,
+    get_net_consumption_sensor_entity_id,
+    get_net_consumption_sensor_name,
+    get_net_consumption_sensor_unique_id,
+    get_next_update_sensor_entity_id,
+    get_next_update_sensor_name,
+    get_next_update_sensor_unique_id,
+    get_plan_explanation_sensor_entity_id,
+    get_plan_explanation_sensor_name,
+    get_plan_explanation_sensor_unique_id,
+    get_read_only_sensor_entity_id,
+    get_read_only_sensor_name,
+    get_read_only_sensor_unique_id,
+    get_recommendation_interval_sensor_entity_id,
+    get_recommendation_interval_sensor_name,
+    get_recommendation_interval_sensor_unique_id,
+    get_solcast_likelihood_selector_entity_id,
+    get_solcast_likelihood_selector_key,
+    get_solcast_likelihood_selector_name,
+    get_solcast_likelihood_selector_unique_id,
+    get_update_interval_sensor_entity_id,
+    get_update_interval_sensor_name,
+    get_update_interval_sensor_unique_id,
+    get_working_mode_sensor_entity_id,
+    get_working_mode_sensor_name,
+    get_working_mode_sensor_unique_id,
+)
+from custom_components.hsem.utils.sensornames.energy import (
+    get_energy_average_sensor_entity_id,
+    get_energy_average_sensor_name,
+    get_energy_average_sensor_unique_id,
+    get_house_consumption_power_sensor_entity_id,
+    get_house_consumption_power_sensor_name,
+    get_house_consumption_power_sensor_unique_id,
+    get_integral_sensor_entity_id,
+    get_integral_sensor_name,
+    get_integral_sensor_unique_id,
+    get_utility_meter_sensor_entity_id,
+    get_utility_meter_sensor_name,
+    get_utility_meter_sensor_unique_id,
+)
+from custom_components.hsem.utils.sensornames.ev import (
+    get_ev_charging_sensor_entity_id,
+    get_ev_charging_sensor_name,
+    get_ev_charging_sensor_unique_id,
+    get_ev_deadline_time_entity_id,
+    get_ev_deadline_time_key,
+    get_ev_deadline_time_name,
+    get_ev_deadline_time_unique_id,
+    get_ev_force_charge_now_switch_entity_id,
+    get_ev_force_charge_now_switch_key,
+    get_ev_force_charge_now_switch_name,
+    get_ev_force_charge_now_switch_unique_id,
+    get_ev_force_discharge_switch_entity_id,
+    get_ev_force_discharge_switch_key,
+    get_ev_force_discharge_switch_name,
+    get_ev_force_discharge_switch_unique_id,
+    get_ev_optimal_charging_plan_sensor_entity_id,
+    get_ev_optimal_charging_plan_sensor_name,
+    get_ev_optimal_charging_plan_sensor_unique_id,
+    get_ev_second_deadline_time_entity_id,
+    get_ev_second_deadline_time_key,
+    get_ev_second_deadline_time_name,
+    get_ev_second_deadline_time_unique_id,
+    get_ev_second_force_charge_now_switch_entity_id,
+    get_ev_second_force_charge_now_switch_key,
+    get_ev_second_force_charge_now_switch_name,
+    get_ev_second_force_charge_now_switch_unique_id,
+    get_ev_second_optimal_charging_plan_sensor_entity_id,
+    get_ev_second_optimal_charging_plan_sensor_name,
+    get_ev_second_optimal_charging_plan_sensor_unique_id,
+    get_ev_second_smart_charging_switch_entity_id,
+    get_ev_second_smart_charging_switch_key,
+    get_ev_second_smart_charging_switch_name,
+    get_ev_second_smart_charging_switch_unique_id,
+    get_ev_second_target_soc_number_entity_id,
+    get_ev_second_target_soc_number_key,
+    get_ev_second_target_soc_number_name,
+    get_ev_second_target_soc_number_unique_id,
+    get_ev_smart_charging_switch_entity_id,
+    get_ev_smart_charging_switch_key,
+    get_ev_smart_charging_switch_name,
+    get_ev_smart_charging_switch_unique_id,
+    get_ev_target_soc_number_entity_id,
+    get_ev_target_soc_number_key,
+    get_ev_target_soc_number_name,
+    get_ev_target_soc_number_unique_id,
+)
+from custom_components.hsem.utils.sensornames.ml import (
+    get_ml_consumption_switch_entity_id,
+    get_ml_consumption_switch_key,
+    get_ml_consumption_switch_unique_id,
+    get_ml_sequential_switch_entity_id,
+    get_ml_sequential_switch_key,
+    get_ml_sequential_switch_unique_id,
+)
+
+__all__ = [
+    # energy
+    "get_integral_sensor_name",
+    "get_integral_sensor_unique_id",
+    "get_integral_sensor_entity_id",
+    "get_energy_average_sensor_name",
+    "get_energy_average_sensor_unique_id",
+    "get_energy_average_sensor_entity_id",
+    "get_utility_meter_sensor_name",
+    "get_utility_meter_sensor_unique_id",
+    "get_utility_meter_sensor_entity_id",
+    "get_house_consumption_power_sensor_name",
+    "get_house_consumption_power_sensor_unique_id",
+    "get_house_consumption_power_sensor_entity_id",
+    # diagnostics
+    "get_working_mode_sensor_name",
+    "get_working_mode_sensor_unique_id",
+    "get_working_mode_sensor_entity_id",
+    "get_degraded_mode_sensor_name",
+    "get_degraded_mode_sensor_unique_id",
+    "get_degraded_mode_sensor_entity_id",
+    "get_read_only_sensor_name",
+    "get_read_only_sensor_unique_id",
+    "get_read_only_sensor_entity_id",
+    "get_next_update_sensor_name",
+    "get_next_update_sensor_unique_id",
+    "get_next_update_sensor_entity_id",
+    "get_missing_entities_sensor_name",
+    "get_missing_entities_sensor_unique_id",
+    "get_missing_entities_sensor_entity_id",
+    "get_hardware_writes_sensor_name",
+    "get_hardware_writes_sensor_unique_id",
+    "get_hardware_writes_sensor_entity_id",
+    "get_net_consumption_sensor_name",
+    "get_net_consumption_sensor_unique_id",
+    "get_net_consumption_sensor_entity_id",
+    "get_force_mode_sensor_name",
+    "get_force_mode_sensor_unique_id",
+    "get_force_mode_sensor_entity_id",
+    "get_update_interval_sensor_name",
+    "get_update_interval_sensor_unique_id",
+    "get_update_interval_sensor_entity_id",
+    "get_last_updated_sensor_name",
+    "get_last_updated_sensor_unique_id",
+    "get_last_updated_sensor_entity_id",
+    "get_battery_soc_sensor_name",
+    "get_battery_soc_sensor_unique_id",
+    "get_battery_soc_sensor_entity_id",
+    "get_recommendation_interval_sensor_name",
+    "get_recommendation_interval_sensor_unique_id",
+    "get_recommendation_interval_sensor_entity_id",
+    "get_plan_explanation_sensor_name",
+    "get_plan_explanation_sensor_unique_id",
+    "get_plan_explanation_sensor_entity_id",
+    "get_force_working_mode_selector_key",
+    "get_force_working_mode_selector_name",
+    "get_force_working_mode_selector_unique_id",
+    "get_force_working_mode_selector_entity_id",
+    "get_solcast_likelihood_selector_key",
+    "get_solcast_likelihood_selector_name",
+    "get_solcast_likelihood_selector_unique_id",
+    "get_solcast_likelihood_selector_entity_id",
+    "get_applier_status_sensor_name",
+    "get_applier_status_sensor_unique_id",
+    "get_applier_status_sensor_entity_id",
+    "get_forecast_accuracy_sensor_name",
+    "get_forecast_accuracy_sensor_unique_id",
+    "get_forecast_accuracy_sensor_entity_id",
+    "get_daily_plan_vs_actual_sensor_key",
+    "get_daily_plan_vs_actual_sensor_name",
+    "get_daily_plan_vs_actual_sensor_unique_id",
+    "get_daily_plan_vs_actual_sensor_entity_id",
+    # ev
+    "get_ev_charging_sensor_name",
+    "get_ev_charging_sensor_unique_id",
+    "get_ev_charging_sensor_entity_id",
+    "get_ev_optimal_charging_plan_sensor_name",
+    "get_ev_optimal_charging_plan_sensor_unique_id",
+    "get_ev_optimal_charging_plan_sensor_entity_id",
+    "get_ev_second_optimal_charging_plan_sensor_name",
+    "get_ev_second_optimal_charging_plan_sensor_unique_id",
+    "get_ev_second_optimal_charging_plan_sensor_entity_id",
+    "get_ev_target_soc_number_key",
+    "get_ev_target_soc_number_name",
+    "get_ev_target_soc_number_unique_id",
+    "get_ev_target_soc_number_entity_id",
+    "get_ev_second_target_soc_number_key",
+    "get_ev_second_target_soc_number_name",
+    "get_ev_second_target_soc_number_unique_id",
+    "get_ev_second_target_soc_number_entity_id",
+    "get_ev_force_discharge_switch_key",
+    "get_ev_force_discharge_switch_name",
+    "get_ev_force_discharge_switch_unique_id",
+    "get_ev_force_discharge_switch_entity_id",
+    "get_ev_smart_charging_switch_key",
+    "get_ev_smart_charging_switch_name",
+    "get_ev_smart_charging_switch_unique_id",
+    "get_ev_smart_charging_switch_entity_id",
+    "get_ev_force_charge_now_switch_key",
+    "get_ev_force_charge_now_switch_name",
+    "get_ev_force_charge_now_switch_unique_id",
+    "get_ev_force_charge_now_switch_entity_id",
+    "get_ev_second_smart_charging_switch_key",
+    "get_ev_second_smart_charging_switch_name",
+    "get_ev_second_smart_charging_switch_unique_id",
+    "get_ev_second_smart_charging_switch_entity_id",
+    "get_ev_second_force_charge_now_switch_key",
+    "get_ev_second_force_charge_now_switch_name",
+    "get_ev_second_force_charge_now_switch_unique_id",
+    "get_ev_second_force_charge_now_switch_entity_id",
+    "get_ev_deadline_time_key",
+    "get_ev_deadline_time_name",
+    "get_ev_deadline_time_unique_id",
+    "get_ev_deadline_time_entity_id",
+    "get_ev_second_deadline_time_key",
+    "get_ev_second_deadline_time_name",
+    "get_ev_second_deadline_time_unique_id",
+    "get_ev_second_deadline_time_entity_id",
+    # controls
+    "get_charge_efficiency_number_key",
+    "get_charge_efficiency_number_name",
+    "get_charge_efficiency_number_unique_id",
+    "get_charge_efficiency_number_entity_id",
+    "get_discharge_efficiency_number_key",
+    "get_discharge_efficiency_number_name",
+    "get_discharge_efficiency_number_unique_id",
+    "get_discharge_efficiency_number_entity_id",
+    "get_read_only_switch_key",
+    "get_read_only_switch_name",
+    "get_read_only_switch_unique_id",
+    "get_read_only_switch_entity_id",
+    "get_extended_attributes_switch_key",
+    "get_extended_attributes_switch_name",
+    "get_extended_attributes_switch_unique_id",
+    "get_extended_attributes_switch_entity_id",
+    "get_verbose_logging_switch_key",
+    "get_verbose_logging_switch_name",
+    "get_verbose_logging_switch_unique_id",
+    "get_verbose_logging_switch_entity_id",
+    "get_batteries_schedule_1_switch_key",
+    "get_batteries_schedule_1_switch_name",
+    "get_batteries_schedule_1_switch_unique_id",
+    "get_batteries_schedule_1_switch_entity_id",
+    "get_batteries_schedule_2_switch_key",
+    "get_batteries_schedule_2_switch_name",
+    "get_batteries_schedule_2_switch_unique_id",
+    "get_batteries_schedule_2_switch_entity_id",
+    "get_batteries_schedule_3_switch_key",
+    "get_batteries_schedule_3_switch_name",
+    "get_batteries_schedule_3_switch_unique_id",
+    "get_batteries_schedule_3_switch_entity_id",
+    "get_schedule_1_start_time_key",
+    "get_schedule_1_start_time_name",
+    "get_schedule_1_start_time_unique_id",
+    "get_schedule_1_start_time_entity_id",
+    "get_schedule_1_end_time_key",
+    "get_schedule_1_end_time_name",
+    "get_schedule_1_end_time_unique_id",
+    "get_schedule_1_end_time_entity_id",
+    "get_schedule_2_start_time_key",
+    "get_schedule_2_start_time_name",
+    "get_schedule_2_start_time_unique_id",
+    "get_schedule_2_start_time_entity_id",
+    "get_schedule_2_end_time_key",
+    "get_schedule_2_end_time_name",
+    "get_schedule_2_end_time_unique_id",
+    "get_schedule_2_end_time_entity_id",
+    "get_schedule_3_start_time_key",
+    "get_schedule_3_start_time_name",
+    "get_schedule_3_start_time_unique_id",
+    "get_schedule_3_start_time_entity_id",
+    "get_schedule_3_end_time_key",
+    "get_schedule_3_end_time_name",
+    "get_schedule_3_end_time_unique_id",
+    "get_schedule_3_end_time_entity_id",
+    # ml
+    "get_ml_consumption_switch_key",
+    "get_ml_consumption_switch_unique_id",
+    "get_ml_consumption_switch_entity_id",
+    "get_ml_sequential_switch_key",
+    "get_ml_sequential_switch_unique_id",
+    "get_ml_sequential_switch_entity_id",
+]
