+++ tests/test_midnight_rollover.py
@@ -16,7 +16,7 @@ Windows.
 import asyncio
 from datetime import UTC, datetime, time, timedelta
 
-from custom_components.hsem.utils.misc import (
+from custom_components.hsem.utils.time_windows import (
     interval_ends_before_window_start,
     is_time_in_window,
 )
